backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib[bcrypt]
python-jose
python-multipart
psycopg2-binary
python-dotenv
python-multipart